Remove magic of gathering of new translation via comment
I remember I added this feature to improve the user experience when working
with the default behavior of Interface Builder, but I don't remember the details.
Also this feature was never documented or tested, so I think it can be removed.

Fingers crossed that there will be no side effects and no one is relying on this.